// src/generated/models/opportunityOut.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * iPhande API
 * Visibility, opportunity continuity, replay, and grace reflection platform.
 * OpenAPI spec version: 0.1.0
 */
import type { OpportunityOutDescription } from './opportunityOutDescription';
import type { OpportunityOutProvince } from './opportunityOutProvince';
import type { OpportunityOutTownOrCity } from './opportunityOutTownOrCity';
import type { OpportunityOutSuburbOrArea } from './opportunityOutSuburbOrArea';
import type { OpportunityOutPlaceCode } from './opportunityOutPlaceCode';
import type { OpportunityOutLatitude } from './opportunityOutLatitude';
import type { OpportunityOutLongitude } from './opportunityOutLongitude';
import type { OpportunityOutCategoryKey } from './opportunityOutCategoryKey';
import type { OpportunityOutServiceNeeded } from './opportunityOutServiceNeeded';
import type { OpportunityOutBudgetAmount } from './opportunityOutBudgetAmount';
import type { OpportunityOutContactName } from './opportunityOutContactName';
import type { OpportunityOutContactPhone } from './opportunityOutContactPhone';
import type { OpportunityOutImageUrl1 } from './opportunityOutImageUrl1';
import type { OpportunityOutImageUrl2 } from './opportunityOutImageUrl2';
import type { OpportunityOutExpiryDate } from './opportunityOutExpiryDate';

export interface OpportunityOut {
  id: string;
  created_by_profile_id: string;
  title: string;
  description: OpportunityOutDescription;
  status: string;
  province: OpportunityOutProvince;
  town_or_city: OpportunityOutTownOrCity;
  suburb_or_area: OpportunityOutSuburbOrArea;
  place_code: OpportunityOutPlaceCode;
  latitude: OpportunityOutLatitude;
  longitude: OpportunityOutLongitude;
  category_key: OpportunityOutCategoryKey;
  service_needed: OpportunityOutServiceNeeded;
  budget_amount: OpportunityOutBudgetAmount;
  contact_name: OpportunityOutContactName;
  contact_phone: OpportunityOutContactPhone;
  image_url_1: OpportunityOutImageUrl1;
  image_url_2: OpportunityOutImageUrl2;
  expiry_date: OpportunityOutExpiryDate;
  created_at: string;
}